refactor: Update error responses, decomposed operation functions
